Clockin (#10)
* Restoration and startup

* New element types substantiated

They aren't ready to go yet, but they now exist.

* Template created

Interactable is about as well as it's going to get before being hooked
into the game script.

* Fundamental Properties Defined

I have yet to hook in to the game in any way, but we now have the basic
properties of these interactables.

* Establish game states and realize new problems

The enemy game states are now integral to the music system, and I've
just realized that because all three types of interactables could be
present in a singular object, I need to rethink how I'm going to code
interactables.

* Properties of classes rolled back into superclass

Objects can only hold 1 script, and game objects are capable of
exhibiting properties of all 3 objects. Time to roll it back up.

* Forgot to save the superclass before committing deletes, sorry!

* Only reset the permeable if it's supposed to be resettable.

* Fixing my own state in git.

Hopefully, when I push these changes, I'm not screwing up anything in
main, but I've been dealing with a crazy amount of weirdness coming from
git, so hopefully this creates a resolution.